+++ doc/whats_new.rst
@@ -514,6 +514,12 @@ API changes summary
    - :class:`multiclass.OneVsRestClassifier` now has a ``partial_fit`` method
      only if the underlying estimator does.  By `Andreas Müller`_.
 
+    - Change the default value for ``learning_rate`` in the
+      `sklearn.manifold.TSNE` constructor from 1000 to 200 (with a warning
+      message if not set explicitely).
+      :issue:`8752` by :user:`Fabio Zanini <iosonofabio>` and
+      :user:`Alexander Fabisch <AlexanderFabisch>`
+
    - Estimators with both methods ``decision_function`` and ``predict_proba``
      are now required to have a monotonic relation between them. The
      method ``check_decision_proba_consistency`` has been added in